+++ src/ReadTask.cpp
@@ -58,7 +58,7 @@ bool ReadTask::run()
     // todo: need to make this use O_DIRECT on the IOC side
     ioc->willRead(cmd->filename, cmd->offset, cmd->count);
     int err;
-    while (resp->returnCode < cmd->count)
+    while ((uint) resp->returnCode < cmd->count)
     {
         try
         {
